// src/enums/FinancialReportTypesEnum.ts
export enum FinancialReportTypesEnum {
  FINANCIAL_BCDKT = 'FINANCIAL_BCDKT',
  FINANCIAL_BLCTT = 'FINANCIAL_BLCTT',
  FINANCIAL_BKQKD = 'FINANCIAL_BKQKD',
  FINANCIAL_BCDPSTK = 'FINANCIAL_BCDPSTK',
  FINANCIAL_SCT = 'FINANCIAL_SCT',
  FINANCIAL_TAX_FINALIZATION_TNCN = 'FINANCIAL_TAX_FINALIZATION_TNCN',
  FINANCIAL_TAX_FINALIZATION_TNDN = 'FINANCIAL_TAX_FINALIZATION_TNDN',
  FINANCIAL_STATEMENT_FOOTNOTES = 'FINANCIAL_STATEMENT_FOOTNOTES',
}